requirements.txt
streamlit
pytorch
torchvision
Pillow
openai  == 0.27.8